Fix license in setup.py
While the ldappool project is MPL and the license info is appropriately
associated with Tarek, setup.py is actually not a file written by him -
it is an exact copy of the boilerplate OpenStack setup.py file which is
Apache licensed. There should be no license issues with the project with
the OpenStack setup.py file being in use, as neither license are
incompatible with each other.

Change-Id: I0d334e624afe62a53b3324ee45ca9bc454ae1d4a