Use standard-library logging to record errors
This is a port of a pull request from the old ldappool repo [1]

ldappool frequently catches and ignores transient errors.
There's nothing generally wrong with this -- but if the user
chooses to give the ldappool module a DEBUG priority level,
they really ought to be able to see these. This change makes it
so (and logs final connection errors with the higher ERROR level).

[1] https://github.com/mozilla-services/ldappool/pull/9

Change-Id: I223584393e6010845a5e7eb6aa619a77f6aba84f