Expose SERVER_DOWN if connection fails
Without this change, ldappool might fail to bind to an LDAP server for
any number of reasons, but eats most of them and instead raises its own
not-very-descriptive error. A network failure or the LDAP server being
down is useful information that a user might be able to do something
with (see this related patch[1]) so let's add that to the list of
exceptions ldappool raises directly.

[1] https://review.openstack.org/#/c/390948

Change-Id: Iebb58f9046707a044ed86d8ce940e4c230ffd2aa