Improve connection retry logging
The logging around the connection retry logic logs an info level
message when it fails to connect to the server, but still has retry
attempts left.  Due to the way we increment the counter for the
number of connection attempts, we log this same message when the
final attempt has failed.  There is an error level log message in
the code that is never reached due to this.

This patch changes the location where we increment the connection
attempt counter, which results in the proper error message being
logged when we fail on the final attempt.  The retry wait is also
moved to be after the info level message, as it states that it is
about to wait before retrying.  The current logic of waiting before
logging the message doesn't actually match what the message says.
I also added a debug level log message that indicates what connection
attempt it is about to make to help troubleshoot the retry logic.

Change-Id: Ib05d510b8816b5e0670b92749412378a100667d0