Add bindep.txt
Without this patch, unit test jobs may fail if the base system does not
have the openLDAP development package installed. This patch ensures that
we don't rely on a fallback bindep file that may or may not be defined
or a base image that may or may not have the required packages
installed.

Change-Id: I5607e7c7307687857f254d9c7bd16af6532ad7e0