Follow the PTI for docs
Use sphinx-build instead of the pbr sphinx extention for building docs
as instructed by the PTI[1].

We remove the reference to the ChangeLog file that is usually generated
by pbr and instead refer to the published reno release notes. Also fixes
the header formatting for the index page, as the headers weren't
rendering at all.

[1] https://governance.openstack.org/tc/reference/pti/python.html

Change-Id: I36142eeb1590c551c6df3bd7d0bc7b7a4a8c692f