Remove lower-constraints, uncap PrettyTable
The lower-constraints jobs are timing out, which means we're failing to
find a match for our specified constraints. These jobs are of limited
value so just drop them.

The change to requirements flags the cap we currently enforce on
PrettyTable. This was previously necessary due to requirements enforcing
the same case, but PrettyTable is now maintained as a Jazzband project
[1] and the cap has been removed.

Finally, while we're here, we can cleanup tox.ini somewhat and remove
now unnecessary warnings from the 'requirements.txt' file (this isn't an
issue with pip's new resolver).

[1] https://github.com/jazzband/prettytable

Change-Id: Ib0fad16f3c9b817756996f79c0b0b61f854583a3
Co-authored-by: Stephen Finucane <stephenfin@redhat.com>